+++ lib/rules/no-top-level-side-effect.ts
@@ -78,6 +78,7 @@ export const noTopLevelSideEffect: Rule.RuleModule = {
       IfStatement: ifTopLevelReportWith(context),
       ForStatement: ifTopLevelReportWith(context),
       ForInStatement: ifTopLevelReportWith(context),
+      ForOfStatement: ifTopLevelReportWith(context),
       WhileStatement: ifTopLevelReportWith(context),
       DoWhileStatement: ifTopLevelReportWith(context),
       SwitchStatement: ifTopLevelReportWith(context)