+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- (`16462ae`) Report top-level for-of statements.
 
 ## [0.2.2] - 2022-12-30